Align diagnostics to the same column
# Details

[ISSUE-4](https://github.com/MeanderingProgrammer/py-requirements.nvim/issues/4)

Currently diagnostics default to being written at the end of each line,
so they end up being relatively unaligned, more a stylistic problem.

The fix is a bit non-trivial for now as we cannot use diagnostics
directly to change the position of the virtual text outside of the
nightly release.

To get around this modified the diagnostics to not display virtual text.

Added extmarks which allow for virtual text to be displayed in much the
same way as diagnostics and mapped highlight groups to match.

Added a helper function to get the largest lenth of any module line.