Support auto completion through patterns
# Details

Previously the cmp source would only run if the file name exactly
matched one of the configured patterns.

Changed this to run if any pattern matches.

However the patterns between autocommands and vim regex are not the
same.

To get around this removed the pattern from the autocommands, instead
now the functions check if they're active by matching against the
same condition as nvim-cmp.