[ISSUE-8] Remove pattern from autocommands
# Details

Patterns in the autocommand can contradict the patterns specified by
the user and cause callback not to run, i.e. `requirements.in`.

Remove the pattern from the callback and rely on the active method to
match the filename or not.

Will increase number of events being processed but should not be enough
to negatively impact user.

Add `pattern` reference to README.