Add final_release option & logic
# Details

Currently we always use all available versions from pypi.

However many of these versions are likely not of interest to the user,
such as alpha, beta, and development versions.

To allow all non final release versions to be filtered out have added
the option final_release, this defaults to false such that the default
behavior is backwards compatible.

If set to true all versions that do not meet the semver definition are
filtered out. This is kind of a hack in that pypi does not use semver.
It just so happens to be the case that versions that do not meet the
semver spec are also all of the non final release versions, which
makes the logic easy to implement.