Filter out yanked package versions from pypi response
# Details

Gates the filtering behavior behind the filter.yanked options,
this is set to true by default.

The logic to do the filtering is very rough, but seems to capture
the current cases, such as with pytest 8.1.0.

There is non backwards compatible change for the final_release option
which is now nested inside of a new filter object.