Use state in API to reduce likelihood of passing incorrect arguments
## Details

Currently everywhere we call our API we pass in a filter configuration.

This makes reasoning about state easier but means we need to pass our
filter configuration everywhere we call the versions API.

This is not very ergonimic and is a bit miss-leading as we cache only
based on the name no matter what filter gets passed on on subsequent
calls.

To make the function calling easier use state directly in the API.

This fixes a minor bug we had in actions.lua which was passing an
invalid filter after the last change. Even though this would not
have impacted anyone due to our cache, having the state in the API
reduces the chance for similar bugs going forward.