Use virt_text_win_col to align diagnostics, NOTE: this breaks plugin for users who have not upgraded to 0.10.0
# Details

Prior to the release of 0.10.0 this plugin would set both diagnostics
as well as extmarks to show version information to the user.

This is because diagnostics did not have as many features as extmarks
in particular horizontal alignment.

Now that 0.10.0 is stable, remove extmarks and use diagnostics only.

While we're at it take advantage of some other features that are
now available, like virtual_text prefix and iterators.

Add a health check to inform users of this version requirement.