Remove dependency on plenary, use vim.system directly
# Details

Since this plugin already uses features from Neovim 0.10.0 we
can also take advantage of vim.system.

This API offers some great flexibility and if we needed can
even run async, something that would be more difficult with
plenary.

Fewer dependencies is also a nice benefit.

Missed the dependency on nvim-treesitter in the example config
in the README, adding it for completeness. This is admittedly
less satisfying then just removing a dependency, but nvim-treesitter
was needed in either case, so will take this as a win.