perf: attach change events at buffer level
## Details

The main purpose of this change is to avoid listening to all TextChanged
and InsertLeave events for all buffers. Instead we listen to only the
BufRead event then from there if the buffer matches our conditions we
attach additional auto commands local to that buffer only.

Some other minor refactoring:

- Rename module api -> pypi
- Create new api module which init.lua extends
- Use pcall on 'cmp' module to avoid errors
- Use buffer from event rather than getting current buffer
- Make treesitter queries configurable, not useful for users but nice
  for development