feat!: use in-process language server (lsp) for completions
## Details

Rather than requiring a source to be added to completion engine by user
create an in-process LSP that provides completion capabilities to
whatever engine the user happens to be running.

Idea for this and the implementation came from [saecki/crates.nvim](https://github.com/saecki/crates.nvim)
and seems to have no downside. Can add more capabilities in the future
such as hover docs in the future.

Users who want to continue using the source independently of an LSP will
need to disable the LSP and potentially re-enable the source:

For `nvim-cmp`:

```lua
require('py-requirements').setup({
    enable_cmp = true,
    enable_lsp = false,
})
```

For `blink.cmp`:

```lua
require('py-requirements').setup({
    enable_lsp = false,
})
```